S2D-55 refactor : 유저가 좋아요한 아이템 검색
유저가 좋아요한 아이템 검색 (/app/user/like/item)
- fetch join을 추가하여 쿼리 성능 개선.
- where 절에 user에 대한 조건 추가.
- Swagger 어노테이션 수정
